refactor admin dashboard to use widgets
