+++ app/Filament/Admin/Pages/Dashboard.php
@@ -2,35 +2,12 @@
 
 namespace App\Filament\Admin\Pages;
 
-use App\Filament\Admin\Resources\NodeResource\Pages\CreateNode;
-use App\Filament\Admin\Resources\NodeResource\Pages\ListNodes;
-use App\Models\Egg;
-use App\Models\Node;
-use App\Models\Server;
-use App\Models\User;
 use App\Services\Helpers\SoftwareVersionService;
-use Filament\Actions\CreateAction;
-use Filament\Pages\Page;
+use Filament\Pages\Dashboard as BaseDashboard;
 
-class Dashboard extends Page
+class Dashboard extends BaseDashboard
 {
     protected static ?string $navigationIcon = 'tabler-layout-dashboard';
-
-    protected static string $view = 'filament.pages.dashboard';
-
-    protected ?string $heading = '';
-
-    public function getTitle(): string
-    {
-        return trans('admin/dashboard.title');
-    }
-
-    public static function getNavigationLabel(): string
-    {
-        return trans('admin/dashboard.title');
-    }
-
-    protected static ?string $slug = '/';
 
     private SoftwareVersionService $softwareVersionService;
 
@@ -39,51 +16,18 @@ class Dashboard extends Page
         $this->softwareVersionService = $softwareVersionService;
     }
 
-    public function getViewData(): array
+    public function getColumns(): int
     {
-        return [
-            'inDevelopment' => config('app.version') === 'canary',
-            'version' => $this->softwareVersionService->currentPanelVersion(),
-            'latestVersion' => $this->softwareVersionService->latestPanelVersion(),
-            'isLatest' => $this->softwareVersionService->isLatestPanel(),
-            'eggsCount' => Egg::query()->count(),
-            'nodesList' => ListNodes::getUrl(),
-            'nodesCount' => Node::query()->count(),
-            'serversCount' => Server::query()->count(),
-            'usersCount' => User::query()->count(),
+        return 1;
+    }
 
-            'devActions' => [
-                CreateAction::make()
-                    ->label(trans('admin/dashboard.sections.intro-developers.button_issues'))
-                    ->icon('tabler-brand-github')
-                    ->url('https://github.com/pelican-dev/panel/issues', true),
-            ],
-            'updateActions' => [
-                CreateAction::make()
-                    ->label(trans('admin/dashboard.sections.intro-update-available.heading'))
-                    ->icon('tabler-clipboard-text')
-                    ->url('https://pelican.dev/docs/panel/update', true)
-                    ->color('warning'),
-            ],
-            'nodeActions' => [
-                CreateAction::make()
-                    ->label(trans('admin/dashboard.sections.intro-first-node.button_label'))
-                    ->icon('tabler-server-2')
-                    ->url(CreateNode::getUrl()),
-            ],
-            'supportActions' => [
-                CreateAction::make()
-                    ->label(trans('admin/dashboard.sections.intro-support.button_donate'))
-                    ->icon('tabler-cash')
-                    ->url('https://pelican.dev/donate', true)
-                    ->color('success'),
-            ],
-            'helpActions' => [
-                CreateAction::make()
-                    ->label(trans('admin/dashboard.sections.intro-help.button_docs'))
-                    ->icon('tabler-speedboat')
-                    ->url('https://pelican.dev/docs', true),
-            ],
-        ];
+    public function getHeading(): string
+    {
+        return trans('admin/dashboard.heading');
+    }
+
+    public function getSubheading(): string
+    {
+        return trans('admin/dashboard.version', ['version' => $this->softwareVersionService->currentPanelVersion()]);
     }
 }
